+++ inst/test-db/globals.R
@@ -25,3 +25,30 @@ make_test(title = "future() - non-exported package objects", tags = c("future", 
     stopifnot(identical(v, non_exported))
   }
 })
+
+
+
+## https://github.com/HenrikBengtsson/future.tests/issues/15
+make_test(title = "future() - NSE '...'", tags = c("future", "globals", "..."), {
+  my_fcn <- function(...) {
+    ## Grab '...' into a Globals object
+    globals <- globals::globalsByName("...", envir=environment())
+
+    ## Evaluate an expression with '...' in an environment that does not
+    ## have an '...' object - hence the parent.frame().  This will produce
+    ## an error unless we pass 'globals' which contains '...'
+    
+    f <- future({
+      fcn <- function() sum(...)
+      fcn()
+    }, envir = parent.frame(), globals = globals)
+    v <- value(f)
+    v
+  }
+
+  y <- my_fcn()
+  stopifnot(y == 0L)  ## sum(c()) is integer, but not sure it's guaranteed
+
+  y <- my_fcn(1:10)
+  stopifnot(identical(y, 55L))
+})